+++ CHANGELOG.md
@@ -29,6 +29,8 @@
  - Added `\Wizaplace\SDK\User\User::getCompanyId`
  - Added `\Wizaplace\SDK\User\User::isVendor`
  - Added `\Wizaplace\SDK\Vendor\Order\OrderService::listTaxes` and associated classes
+ - Added `\Wizaplace\SDK\Pim\Category\CategoryService::listCategories`
+ - Added `\Wizaplace\SDK\Pim\Tax\TaxService::listTaxes`
 
 ### Bugfixes